+++ .gitignore
@@ -1,0 +1,3 @@
+bundle.js
+.idea
+node_modules
